requirements.txt
streamlit
numpy
pandas
pillow
scikit-learn
